chore(suite): reorganize component files
