+++ packages/suite/src/components/wallet/CoinjoinSummary/AnonymityChart.tsx
