+++ packages/suite/src/views/wallet/send/components/Options/components/BitcoinOptions/components/CoinControl.tsx
@@ -9,7 +9,7 @@ import { useActions } from '@suite-hooks';
 import { ExtendedMessageDescriptor } from '@suite-types';
 import { Pagination } from '@wallet-components';
 import { Checkbox, Icon, Switch, variables } from '@trezor/components';
-import { UtxoSelectionList } from '@wallet-components/UtxoSelectionList';
+import { UtxoSelectionList } from '@wallet-components/CoinControl/UtxoSelectionList';
 import { useSendFormContext } from '@wallet-hooks';
 import { useBitcoinAmountUnit } from '@wallet-hooks/useBitcoinAmountUnit';
 import { TypedFieldError } from '@wallet-types/form';
